Move api classes to separate kt file
